Get rid of a warning about integer truncation
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@85918 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/peopsxgl/soft.c
@@ -1212,7 +1212,7 @@ static __inline int shl10idiv(int x, int y)
 {
  long long int bi=x;
  bi<<=10;
- return bi/y;
+ return (int)(bi/y);
 }
 
 #endif
